[주문] 데이터 필드 변경 (쿼리dsl 리셋 요망)

+++ javacafe/src/main/java/kr/co/javacafe/dto/ShopJoinDTO.java
@@ -21,7 +21,7 @@ public class ShopJoinDTO {
 	private String rdesc;
 	private String rtext;
 	private int rcost;
-	private int rprice;
+	private Long rprice;
 	private int rkcal;
 	private String rimg;
 	private int rstate;
